Added BMI and BFP Calculator
Body fat percentage calculator added to the webpage.
When the BMI is calculated, it will automatically gets printed in Body fat percentage BMI parts: where id is used to reference the value.
added some JS code for Body Fat Percentage Calculator.
CSS for table and it's properties are added
BFP calculation for male and female need to be added.